Move keys service to its own submodule
