+++ liana-gui/src/installer/message.rs
@@ -52,7 +52,7 @@ pub enum Message {
     MnemonicWord(usize, String),
     ImportMnemonic(bool),
     RedeemNextKey,
-    KeyRedeemed(ProviderKey, Result<(), services::Error>),
+    KeyRedeemed(ProviderKey, Result<(), services::keys::Error>),
     AllKeysRedeemed,
     BackupWallet,
     ExportWallet(Result<String, backup::Error>),
@@ -176,7 +176,7 @@ pub enum ImportKeyModal {
     NameEdited(String),
     ManuallyImportXpub,
     ConfirmXpub,
-    UseToken(services::api::KeyKind),
+    UseToken(services::keys::api::KeyKind),
     TokenEdited(String),
     ConfirmToken,
     SelectKey(usize),
